refactor(ai-agent): move browse_workspace into AgentService
